+++ binder-epub/src/main/java/net/kemitix/binder/epub/MarkdownEpubRenderer.java
@@ -1,7 +1,9 @@
-package net.kemitix.binder.epub.mdconvert;
+package net.kemitix.binder.epub;
 
 import coza.opencollab.epub.creator.model.Content;
 import net.kemitix.binder.epub.EpubRenderer;
+import net.kemitix.binder.epub.mdconvert.Epub;
+import net.kemitix.binder.epub.mdconvert.FootnoteGenerator;
 import net.kemitix.binder.markdown.Context;
 import net.kemitix.binder.markdown.MarkdownConverter;
 import net.kemitix.binder.spi.HtmlSection;
